backend/requirements.txt
django>=5.1
djangorestframework
psycopg2-binary
django-cors-headers
pyjwt
python-dotenv
